Minor cleanup for simulation widget
